refactor: type 파일 분리. children props 상속 구조 통일
