+++ src/context/item/provider.tsx
@@ -3,9 +3,9 @@ import { generateItems } from "../../utils";
 import { ItemContext } from ".";
 import { useCallback, useMemo } from "../../@lib";
 
-export const ItemProvider: React.FC<{
-  children: React.ReactNode;
-}> = ({ children }) => {
+export const ItemProvider: React.FC<{ children: React.ReactNode }> = ({
+  children,
+}) => {
   const [items, setItems] = useState(generateItems(1000));
 
   const addItems = useCallback(() => {
